chore: Update Angular and Bootstrap dependencies
This commit updates the Angular and Bootstrap dependencies in the project. It upgrades Angular to version 17.3.0 and adds Bootstrap version 5.3.3 as a new dependency. This update ensures compatibility with the latest versions of these libraries and allows us to take advantage of any bug fixes or new features they provide.